sparated market_data into 3 files an dcreated the 2 test for marketdata and stock data
