+++ CMakeLists.txt
@@ -14,7 +14,7 @@ include_directories(include)
 
 # Add the main executable
 add_executable(Option_pricer main.cpp
-        src/market_data.cpp
+        src/market_data/market_data.cpp
         src/option/options/base_option.cpp
         src/option/options/single_path/base_single_path_option.cpp
         src/option/options/single_path/european_option.cpp
@@ -41,6 +41,8 @@ add_executable(Option_pricer main.cpp
         src/payoff/single_strike/digital/payoff_digital_put.cpp
         src/payoff/double_strikes/base_payoff_double_strikes.cpp
         src/payoff/double_strikes/payoff_double_digital.cpp
+        src/market_data/market_data_observer.cpp
+        src/market_data/stock_data.cpp
 )
 
 # Enable testing
