+++ tests/integration/CMakeLists.txt
@@ -4,7 +4,7 @@ find_package(GTest REQUIRED)
 add_executable(Integration_tests
         ${CMAKE_CURRENT_SOURCE_DIR}/../test_main.cpp
         test_integration_option.cpp
-        ${CMAKE_CURRENT_SOURCE_DIR}/../../src/market_data.cpp
+        ../../src/market_data/market_data.cpp
         ${CMAKE_CURRENT_SOURCE_DIR}/../../src/option.txt
         ${CMAKE_CURRENT_SOURCE_DIR}/../../src/payoff.txt
         ../config/mock_classes.h
