+++ tests/unit/CMakeLists.txt
@@ -3,7 +3,7 @@ find_package(GTest REQUIRED)
 
 # Create the executable called Unit_tests
 add_executable(Unit_tests ${CMAKE_CURRENT_SOURCE_DIR}/../test_main.cpp
-        test_market_data.cpp
+        market_data/test_market_data.cpp
         test_option.cpp
         payoff/single_strike/vanilla/test_payoff_vanilla_call.cpp
         payoff/single_strike/vanilla/test_payoff_vanilla_put.cpp
@@ -11,7 +11,7 @@ add_executable(Unit_tests ${CMAKE_CURRENT_SOURCE_DIR}/../test_main.cpp
         payoff/single_strike/digital/test_payoff_digital_put.cpp
         payoff/double_strikes/test_payoff_double_digital.cpp
         # Add the necessary source files
-        ../../src/market_data.cpp
+        ../../src/market_data/market_data.cpp
         ../../src/option/options/base_option.cpp
         ../../src/option/options/single_path/base_single_path_option.cpp
         ../../src/option/options/single_path/european_option.cpp
@@ -38,6 +38,7 @@ add_executable(Unit_tests ${CMAKE_CURRENT_SOURCE_DIR}/../test_main.cpp
         ../../src/payoff/single_strike/digital/payoff_digital_put.cpp
         ../../src/payoff/double_strikes/base_payoff_double_strikes.cpp
         ../../src/payoff/double_strikes/payoff_double_digital.cpp
+        market_data/test_stock_data.cpp
 )
 
 target_link_libraries(Unit_tests GTest::GTest GTest::Main GTest::gmock_main)
