+++ tests/unit/market_data/test_market_data.cpp
@@ -1,7 +1,7 @@
 #include <gtest/gtest.h>
 #include <memory>
-#include "market_data.h"
-#include "../config/test_support.h"
+#include "market_data/market_data.h"
+#include "../../config/test_support.h"
 
 // Test fixture for MarketData tests
 class MarketDataTest : public ::testing::Test {
@@ -93,14 +93,4 @@ TEST_F(MarketDataTest, RiskFreeRate) {
 
     marketData->setR(0.03);
     EXPECT_DOUBLE_EQ(marketData->getR(), 0.03);
-}
-
-
-class StockDataTest : public ::testing::Test {};
-
-TEST_F(StockDataTest, InvalidStockData) {
-    EXPECT_THROW(StockData(-150.0, 0.2, 0.01), std::invalid_argument);
-    EXPECT_THROW(StockData(150.0, -0.2, 0.01), std::invalid_argument);
-    EXPECT_THROW(StockData(150.0, 0.2, -0.01), std::invalid_argument);
-    EXPECT_THROW(StockData(150.0, 0.2, 1.5), std::invalid_argument);
 }
