+++ tests/unit/market_data/test_stock_data.cpp
@@ -1,0 +1,10 @@
+#include "gtest/gtest.h"
+#include "market_data/stock_data.h"
+class StockDataTest : public ::testing::Test {};
+
+TEST_F(StockDataTest, InvalidStockData) {
+    EXPECT_THROW(StockData(-150.0, 0.2, 0.01), std::invalid_argument);
+    EXPECT_THROW(StockData(150.0, -0.2, 0.01), std::invalid_argument);
+    EXPECT_THROW(StockData(150.0, 0.2, -0.01), std::invalid_argument);
+    EXPECT_THROW(StockData(150.0, 0.2, 1.5), std::invalid_argument);
+}
